Created Dummy Search blueprint. Moved blueprints to folder.
